Fairly major refactor of CLI
